added production grade security

+++ src/main/java/com/dheerajmehra/hospitality_service/Service/HotelServiceImpl.java
@@ -5,12 +5,15 @@ import com.dheerajmehra.hospitality_service.dto.HotelInfoDto;
 import com.dheerajmehra.hospitality_service.dto.RoomDto;
 import com.dheerajmehra.hospitality_service.entity.Hotel;
 import com.dheerajmehra.hospitality_service.entity.Room;
+import com.dheerajmehra.hospitality_service.entity.User;
 import com.dheerajmehra.hospitality_service.exception.ResourceNotFoundException;
+import com.dheerajmehra.hospitality_service.exception.UnAuthorizedException;
 import com.dheerajmehra.hospitality_service.repository.HotelRepository;
 import com.dheerajmehra.hospitality_service.repository.RoomRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -26,11 +29,18 @@ public class HotelServiceImpl implements HotelService{
     private final InventoryService inventoryService;
     private final RoomRepository roomRepository;
 
+    private boolean isOwner(Hotel hotel){
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        assert user != null;
+        return user.equals(hotel.getOwner());
+    }
     @Override
     public HotelDto createNewHotel(HotelDto hotelDto) {
         log.info("Creating a new hotel with name: {}", hotelDto.getName());
         Hotel hotel = modelMapper.map(hotelDto, Hotel.class);
         hotel.setActive(false);
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        hotel.setOwner(user);
         hotel = hotelRepository.save(hotel);
         log.info("Created a new hotel with ID: {}", hotelDto.getId());
         return modelMapper.map(hotel, HotelDto.class);
@@ -42,6 +52,8 @@ public class HotelServiceImpl implements HotelService{
         Hotel hotel = hotelRepository
                 .findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: "+id));
+        if(!isOwner(hotel))
+            throw new UnAuthorizedException("user is not authorize to perform this action");
         return modelMapper.map(hotel, HotelDto.class);
     }
 
@@ -53,6 +65,8 @@ public class HotelServiceImpl implements HotelService{
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: "+id));
         modelMapper.map(hotelDto, hotel);
         hotel.setId(id);
+        if(!isOwner(hotel))
+            throw new UnAuthorizedException("user is not authorize to perform this action");
         hotel = hotelRepository.save(hotel);
         return modelMapper.map(hotel, HotelDto.class);
     }
@@ -63,7 +77,8 @@ public class HotelServiceImpl implements HotelService{
         Hotel hotel = hotelRepository
                 .findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: "+id));
-
+        if(!isOwner(hotel))
+            throw new UnAuthorizedException("user is not authorize to perform this action");
         for(Room room: hotel.getRooms()) {
             inventoryService.deleteAllInventories(room);
             roomRepository.deleteById(room.getId());
@@ -78,7 +93,8 @@ public class HotelServiceImpl implements HotelService{
         Hotel hotel = hotelRepository
                 .findById(hotelId)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: "+hotelId));
-
+        if(!isOwner(hotel))
+            throw new UnAuthorizedException("user is not authorize to perform this action");
         hotel.setActive(true);
 
         // assuming only do it once
